backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and vector operations
chromadb==0.4.18
neo4j==5.14.1
pydantic==2.5.0

# AI and ML libraries
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3

# Text processing
spacy==3.7.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.11.0
isort==5.12.0 